requirements.txt
# ── Core Data Science ─────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
statsmodels==0.14.2

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
imbalanced-learn==0.12.3   # SMOTE for class imbalance

# ── Model Interpretation ──────────────────────────────────────────────────────
shap==0.46.0

# ── MLOps ─────────────────────────────────────────────────────────────────────
mlflow==2.14.3

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.3
pydantic==2.8.2

# ── Visualization ─────────────────────────────────────────────────────────────
plotly==5.22.0
matplotlib==3.9.1
seaborn==0.13.2
streamlit==1.36.0

# ── GenAI / RAG ───────────────────────────────────────────────────────────────
llama-index==0.11.0
llama-index-llms-openai==0.2.2
llama-index-embeddings-huggingface==0.3.0
openai==1.35.14

# ── Config & Utilities ────────────────────────────────────────────────────────
pyyaml==6.0.2
python-dotenv==1.0.1
joblib==1.4.2
tqdm==4.66.4

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-cov==5.0.0
httpx==0.27.0              # async test client for FastAPI

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.5
nbformat==5.10.4